+++ tests/unit/test_envelopes.c
@@ -15,7 +15,7 @@
 static char *const SERIALIZED_ENVELOPE_STR
     = "{\"dsn\":\"https://foo@sentry.invalid/42\","
       "\"event_id\":\"c993afb6-b4ac-48a6-b61b-2558e601d65d\",\"trace\":{"
-      "\"public_key\":\"foo\",\"org_id\":\"\",\"sample_rate\":0,\"sample_"
+      "\"public_key\":\"foo\",\"sample_rate\":0,\"sample_"
       "rand\":0.01006918276309107,\"release\":\"test-release\",\"environment\":"
       "\"production\",\"sampled\":\"false\"}}\n"
       "{\"type\":\"event\",\"length\":71}\n"